Fix invalid MySQL syntax in Flyway patch
Flyway recently (v5.2.1) switched from the MariaDB driver to MySQL.
Unlike MariaDB, MySQL has a strict requirement that comment lines
beginning with "--" have at least one whitespace after them before
the comment begins.

One of our patches was missing this whitespace, and needs to be
updated for compliance with the MySQL requirement. The checksum
for any previous applications of this patch will need to be updated.
(The new checksum is -884764634.)