Dockerfile: remove unused tar artifacts
When untarring the php source tarball to get the php.ini-production
file, we were leaving the entire php source in the image, which was
adding about 13MB to that layer.

One option would be to simply delete /usr/src/php after copying the
php.ini-production file, but even more efficient is to only untar
that single file. This reduces the final image size by ~13MB.

A couple notes:

* We use "-O" to output the file to stdout and then redirect that
  to our real php.ini file so that we don't have to perform the
  extra step of copying to file.

* We use "--wildcards" so that we don't need to know the exact
  version of php we're using in order to extract the file, since
  the root level directory inside the tarball is "php-7.2.2" (or
  whatever version happens to be in use). There should be no other
  root level directories except that one, so it is safe to use the
  "*/php.ini-production" wildcard.