Dockerfile: improve tar command
Remove the "--wildcards */php.ini-production" with using the PHP
version explicitly "php-$PHP_VERSION/php.ini-production".

This is facilitated by the fact that the base php image defines
the variable $PHP_VERSION, so there is no need to modify the
Dockerfile locally when the version changes (as I suggested in
the previous commit message).