index.php: require login.php
Allow the login page to display from index.php without a 302
redirection.

When looking at the SEO rankings of www.smrealms.de, we see
warnings because of the 302 redirect from index.php to login.php,
which registers it (unfavorably) as a "temporary" page.

While this does allow for a consistent display of login.php in
the browser URL when on the login page, I think the performance
gain (one less round trip to the server) of displaying the login
page for "www.smrealms.de" is worth any inconsistency we may see
in the full URL.