Replace escape_string with escapeString
Finally remove obsolete method MySqlDatabase::escape_string.

Fixes the following bugs:

* If a player had an IRC Nick or a HoF name that included characters
  escaped by `htmlentities`, then other players could switch to that
  nick/name in Preferences without getting the "name already in use"
  error, since these fields are not escaped in the database.

* Handle player name escaping consistently. Since "Join Game" escapes
  html entities, we need to also do the same for editing player name
  in all other places (both "Edit Account" and "Preferences").
  This appears to be the only field that is escaped pre-database.

It also illuminated a lot of places that are vulnerable to either
a) XSS attacks when user data is displayed in html or b) don't
display properly because html entities aren't escaped. I will
need to go through and try to handle this in a more consistent way
in the future. It is not clear yet if this should be done before
putting the data in the database, or upon display (probably the
latter so that we preserve the raw input).