Do not auto-link social logins
When using the social login feature, if the registered e-mail address
was already associated with an SMR account, it would automatically
link to that account without any extra validation.

For added security, require the user to enter their account password
before creating the link (i.e. it sends them to the "Link Existing
Login" page, but without the option to create a new account).

The matching login name will be auto-populated, but it can be changed
should the user wish to link to some other account (e.g. their multi).
This doesn't cause any "duplicate e-mail in database" problems if
they choose to link a different account since linking doesn't use the
social e-mail for anything else.