council.inc: fix how President is chosen
The President must be one of the council members, otherwise there
is no President.

This issue was introduced in 75772c1846d, which caused the President
to be chosen as the first player (in descending experience) to have
a qualifying alignment, even if they weren't in the council.

This fix is a slight performance regression.