Remove unnecessary uses of the URL constant
All of the internal redirection can be done relative to the domain
root. The URL will now only be used for:

* social login callback URLs
* links back to the server in e-mail messages

As a result, we no longer need to have the correct URL for basic
game functionality (which is useful when debugging or migrating).
However, it does make it less obvious when the *wrong* URL is used.