Add container_name to smr service
Since we're using a common block for both the `smr` and `smr-dev`
services, we need to use the same container name for both so that
traefik knows how to communicate with it. This can be achieved by
setting a container name in the common block.

We add a SMR_HOST environment variable for this so that we can
have different host names for each instance.