SmrAlliance: prohibit joining disbanded alliances
In the `canJoinAlliance` method, check if the alliance has disbanded,
and if so, prohibit joining for that reason.

This fixes an issue where people were extended alliance invitations
to an alliance that afterwards disbanded. Accepting the invitation
would trigger a database error trying to send a join message to
player ID 0.