Use class autoloader
Create a new function `get_class_loc`, which we will register
as the autoloading function for classes. With this change, we no
longer need to explicitly `require` files with class definitions.

We move the autoloading functions out of Globals.class.inc into
a separate (minimal) autoloader.inc, which can be included before
the autoloader is registered with the fewest number of dependencies.

Another benefit is that the SmrSession::init function will only be
called if the SmrSession is needed. (There were a lot of tools that
don't need a session, but were instantiating one anyway since they
used the same loading scripts as the main engine.)

NOTE: we need to be super careful about what classes we instantiate,
and when, to avoid issues with the autoloader defining a class before
we have a chance to set the `$overrideGameID` (which determines the
subdirectory in which to preferentially look for classes).

Refactor some of the config files for a better ordering of operations.
The `config.php` file is now *exclusively* constant definitions, and
`config.inc` now does all the autoloading and template initialization.