Migrate away from anonymous data volumes
For data that we care about (in this case, the game database
and the player uploads), it is easier to manage a local directory
than an anonymous data volume. In particular, back up and restore
can be done without interacting with docker (if desired), and we
have more control over where the data is located (e.g. if we want
it on a raided partition instead of confined to wherever docker
puts the data volumes in the OS partition).