login_processing.php: save the SmrSession sooner
There was a lot of other processing between when the SmrSession
was created during login and when it was actually saved to the
database. The longer this time interval, the greater the opportunity
for a race condition.

Save the SmrSession as soon as possible (immediately after it is
updated with the account_id and the next page SN. This is not
expected to fix the problem, but may reduce its frequency.

See #635 for more detail.