Handle microtime as a float instead of a string
This allows us to simplify most of the code that handles microtimes,
since we no longer need to do string parsing to calculate time diffs.

We remove the following three functions:

* microtimeSec: replaced by `intval($microtime)`
* microtimeMsec: unused
* microtimeDiff: replaced by subtraction

The only remaining complication is the storage of microtimes as a
custom-formatted string in the `account_has_logs` table. We handle
this for now by converting the float microtime into the appropriate
string format, but we may want to consider re-typing the `microtime`
column to some datetime or float format in the future.