SmrPlanet: explicitly include the SmrPlanetType file
There are times when the derived planet type classes need to be
loaded even when the base class isn't (i.e. when we deserialize
a planet combat log).

Since the SmrPlanetType file defines multiple classes, it is not
handled by the autoloader, so we need to include it explicitly.

This should fix a PHP Fatal Error that was being encountered
during planet busts / viewing combat logs.