Remove hacks for bar pages
Containers will now directly point to the bar page they intend
to go to instead of using the `script` session variable hack in
`bar_main.php`. The common processing previously performed in
`bar_main.php` was not actually necessary, so we removed it.
Then, for clarity, `bar_opening.php` was moved to `bar_main.php`.

This brings the bar pages in line with all the other multi-page
setups (planets, locations, banks, etc.), and allows us to remove
special cases for bar pages because of this old, hacky setup.

Additional changes:

* Use the `LocationID` var provided when entering the bar and
  pass that to each of the containers so that we can correctly
  identify what location we're in at all times (e.g. when there
  are two bars in the same sector).

* Removed a bunch of the "getBarXHREF" methods from Globals because
  they were only used once (except for the Bar menu pages).

* Cleaned up some unnecessary session variables (either where they
  were unused or by breaking a multi-purpose page into multiple
  pages).

* Renamed some pages for clarity.
  - bar_gambling_lotto.php -> bar_lotto_buy.php
  - bar_gambling_processing.php (lotto) -> bar_lotto_buy_processing.php