Template: do not allow re-assigning variables
This establishes some sense of "constness" for template variables.
In general, we do not want to accidentally override these variables,
especially for those that have a special meaning (e.g. `GameID`,
`AccountID`, and `ThisPlayer` among many others).

There were a handful of cases where we were overwriting variables,
but all legitimate uses so far can be modified to only set it once.

* Move some of the template assignments in config.inc to `do_voodoo`
  in smr.inc. This does require adding some template parameters to
  pages that use the Template class, but don't use `do_voodoo`
  (i.e. that call `Template::display` explicitly).

* Rework some of the logic for setting the ajax refresh time so
  that it is consolidated into a single location.