Template: modify `display` interface
Replace the argument `forAjax` with `outputXml`, which now only
serves the role of deciding whether or not to serve the content
as XML or HTML.

Remove the unused argument `xmlResponse`, which was replaced a
long time ago with the 'AJAX_ENABLE_REFRESH' template variable.
We now make this variable optional, and disable all ajax work
if it is false or undefined (i.e. no ajax output even if `outputXml`
is true).