Fix issue with autoloader and viewing history databases
Because of the automatic way that we set `OverrideGameID` (based
on a `game_id` or `GameID` in the container), we were inadvertently
trying to autoload classes based on the game type associated with
the *historical* game ID.

This doesn't make sense because the historical games are in a
separate database, and we don't have game types for those.

Worse, if the game ID of the historical game doesn't exist in the
current database (as is the case on the live server), then the user
would get an unreportable database error.

To fix this, we need to avoid using the special container variable
name `game_id` (so we replace it with `view_game_id`).