smr.inc: simplify logoff infrastructure
Do not use a separate `logoff_preprocessing.php` workaround.
This "page" (which didn't actually exist) was used to prevent some
of the `do_voodoo` processing from happening so that it could do
and early redirect to `logoff.php`. However, it is much simpler to
just skip this preprocessing step and link directly to `logoff.php`.

This change is facilitated in part by 5bbd4d5a97, which turned
`logoff.php` into a processing page instead of a display page.