Prevent mysql password leaking in backtraces
The mysqli database connector provides a (relatively) secure error
message when a database connection fails, e.g.,

Connection failed: Access denied for user 'smr'@'XXX.XX.X.X' (using password: YES)

and does not directly throw an exception, which would display the
password passed to its constructor (since PHP backtraces will print
function arguments in plain text). However, since our function
`MySqlDatabase::error` takes the mysqli error message and packages
it into an exception, the password is leaked by the `MySqlDatabase`
constructor line of the resulting backtrace. But we want to be able
to raise an exception (securely) since that's how we perform our
standard error handling.

Fortunately, we can plug this leak by loading the password directly
in MySqlDatabase instead of the derived classes (so it doesn't need
to be passed to any function other than `mysqli`, which is outside
the scope of the backtrace).

This does somewhat contaminate the SMR-agnostic design of the
MySqlDatabase base class, but it is still effectively generic, and
it is well worth the added security.

NOTE: Our solution here would not have worked if we were using PDO
instead of mysqli, because PDO prints the database password in the
connection error message (because it internally raises an exception).
It seems that people recommend the very unsatisfying workaround of
catching the exception and then filtering the password out of the
error message string.