Move setupMailer from smr.inc to config.inc
Since `logException` may need to send mail, and `smr.inc` may not
be loaded at the time an exception is thrown, we want to make sure
that the `setupMailer` function is available in the same file that
tries to send bug report e-mails.

This fixes a specific issue of `login_social_processing.php` not
mailing error reports.