Create a new Template for each page
If we initialize the Template at the config.inc level, then it will
be reused between pages that call `forward`. In general, this is not
a problem, because we typically don't set template variables until
we're on the final page that we intend to display. However, it has
caused a problem in at least one case, where we forwarded to the
error.php page after the `PageTopic` variable had already been set
(thus triggering an Exception).

This particular issue could have been obviated by not setting the
`PageTopic` template variable until after all `create_error` sanity
checks had been performed on the display page. But this is not a
safe assumption (it depends on the coder always remembering when to
assign to the template), nor does it fix any cases that we haven't
yet encountered.

Instead, creating the template in `do_voodoo` ensures that we have
a fresh template for each page, guaranteed to be free of any stale
template variables.

(This does require that we also create a Template on any page that
doesn't use `do_voodoo`, e.g., the Galaxy Map and the login page.)