account_edit.php: only search explicitly
The query to find an account was being performed any time the page
was loaded, not just when clicking the "Search" button. If any of
the accounts match the default search fields (empty strings), then
a results page would be displayed without ever showing the search
form.

We can fix this by checking to see if the "Search" button has been
clicked, and only then perform the search.

Added an error message if the search yields no results.