Template.class.inc: add an error-checking dtor
There could possibly be some historical cases where template
variables are intentionally assigned in processing pages to be
displayed on the next page.

This usage would be broken by #658, where we create a new
Template for each page.

We want to make sure there are no legitimate cases in use, so
we will now warn in the destructor of Template if `display` was
not called _and_ if the `data` array is not empty (i.e. a variable
was assigned but not displayed).

This warning has no traceability, but see #658 for (expensive)
debugging options.

Thanks to @Page- for the suggestion!