Don't assign to template before forwarding
There were a handful of files that were assigning template variables
prior to forwarding (either through `forward` or `create_error`),
and in the "template for each page" framework, this would currently
generate a warning.

Though these issues were innocuous (as opposed to an important message
being assigned to a template parameter only to be thrown away after
forwarding), we want to avoid these situations as much as possible.