Remove tools/cache_player.php
This ancient script is no longer functional (refers to files and
database fields that no longer exist). It does not provide any
code that is relevant anymore, even for future use.

It appears to have originally been used for some database
maintenance, but then was co-opted to send news to IRC.