Remove blackjack database table
This table was only used to store the previous hand, which was then
compared to the current hand and raised an error if they were the
same. I was unable to trigger this error, and I now believe it is a
relic of a very old implementation of the forwarding mechanism
(which may have allowed people to manipulate their hands).

In the current forwarding mechanism, the blackjack hand is entirely
server-side, so it should not be possible to manipulate the cards.
And since the table stores the hand as an object, keeping it prevents
us from updating the type of object the cards are stored in.

So we remove this (almost certainly) obsolete error check and also
the `blackjack` database table.