Make sure & is encoded in displayed VoteSite URLs
If the `&` characters in the voting site URLs are not encoded, then
`DOMDocument::loadHTML` will output a htmlParseEntityRef warning due
to a missing `;`, since this is how entities are encoded in HTML
(e.g. `&amp;`). Use `htmlspecialchars` to make the sure `&` characters
are properly encoded.

We currently suppress this error message in the Template class, but
this brings us one step closer to not needing to suppress it.