Do not use InputFields as an HTML element id
We should only use `InputFields` as an element `class`, and never
an `id`. The reason is because the `id` must be unique on a page,
and the CSS for `InputFields` is intended to style all input
elements.

This fixes the `DOMDocument::loadHTML` warning:

> ID InputFields already defined

It also fixes a similar warning in HTML validators.