Remove obsolete align attribute
The `align` attribute is not valid in HTML5 and needs to be replaced.
This is mostly a simple change, either removing `align="center"`
altogether, or replacing it with `class="center"`.

Since each page needs to be individually checked for more complex
CSS interactions, the changes are not always trivial. Some formatting
changes were made, but they are not independently identified here.

A few special notes:

* Table headers do not need to be centered, because the CSS does
  that automatically for all tables.

* When a table is included in a (align)-centered `<div>`, it is
  also centered, but the text inside the table is not. In this case,
  we want to use `centered` instead of `center` as the table class.

* For `<img>` tags with `align`, we generally want to use either
  `class="bottom"` (for making it align better with text) or
  `style="float: right;"` to position it in the same way on the page.

NOTE: also fixed a small number of unrelated issues:

* PlanetList{,Financial}.inc: remove duplicate `id` elements
* Modified spacing and alignment to visual taste
* Replace some HTML styles with CSS classes (e.g. bold and big)
* combat_log_list.php: make table `inset`