hall_of_fame_player_detail.php: synchronize
There is duplicated code in this script with `hall_of_fame_new.php`.
This should be fixed. However, in the meantime, we want to bring over
a change that was made to it a while ago to re-synchronize these
code blocks.

From 8c76ab4e7f1d1d:

HOF: make categories link to the subcategory

Currently, if you want to see the "HoF -> Planets" page, you have
to click on the "Buildings" button in the "Planets" subcategory,
and then click "Planets" in the breadcrumb you get on that page.

This Category link allows you to go directly to the subcategory
page.