Modify ideal trade prices
Reorganize the SmrPort::getIdealPrice method for clearer presentation
of the magnitude of each effect (relations / distance / supply).

Modify the formulas in the following ways:

**DISTANCE**

Ideal price grows more slowly with distance (and scales the same
with distance for both buy & sell). This was needed because the
sell prices for jump routes (>10 sectors) were too lucrative.

**RELATIONS**

Reduce the penalty for having less than perfect relations. This
effect was way too severe for relations below zero.

BUGFIX: The price to buy goods is now correctly scaled with relations.
Previously, this was mostly backwards -- you'd get a cheaper price
the worse your relations were (due to a perhaps accidental quadratic
term in the equation). This makes low-relations purchases cost more.

**PORT LEVEL**

Remove the port level factor in the buying formula. This was an
unnecessary complication that needlessly punished trading at lower
level ports (which was already discouraged due to not having high
tier goods). It was also never a factor in the selling formula.

**SUPPLY**

Flip effect of demand on the sell price so that it consistently
follows the law of supply and demand (buy price decreases with supply
and sell price increases with demand).

NOTE: This has a fairly major implication for gameplay dynamics, since
it will no longer be the best strategy to drain a route. Instead, you
will be rewarded by keeping demand high, and perhaps trading more than
just one route to maximize profit.