Fix weapon Show/Hide ajax
* Clicking the Show/Hide link no longer stops automatic ajax updates.
* Multiple clicks of Show/Hide on the same page are now respected.

Add `'AJAX'=>true` to the container used for the JavaScript-only
Show/Hide link so that when we run toggle_processing.php,
`USING_AJAX=true`.

Since the Show/Hide HTML changes are performed exclusively in JS,
the only purpose of the call to toggle_processing.php is to update
the `display_weapons` setting for the SmrPlayer. Therefore, we will
do an early exit in toggle_processing.php (when triggered with ajax)
so that the remainder of `do_voodoo` (i.e. template display and
SmrSession updating) does not get run.

The only somewhat backwards aspect of this fix is that we had to make
SmrPlayer::setDisplayWeapons update the `player` database table
directly. This is because ajax triggers, as designed here, do not
acquire a sector lock. (This is partly due to a degeneracy of the
USING_AJAX constant, which disables both SmrSession updates and
sector locks, because normal ajax auto updates don't need either.)
However, we don't want to update the entire `player` row without
a sector lock, since this could clobber other changes to this player,
so we are forced to do a direct update of the `display_weapons` table
column.