Output XML for all ajax template displays
Because USING_AJAX has a lot of side-effects (like not updating
the SmrSession and skipping the sector lock), we now only demand
that ajax queries set _either_ USING_AJAX or AJAX_CONTAINER.

Now, when either of these constants is set to true, we will output
XML from Template::display. We do this by using the `$outputXml`
argument more consistently with Template::display instead of
relying on USING_AJAX directly.