Change how the ajax=1 query parameter is set
We no longer set `&ajax=1` query parameter in SmrSession::getNewHREF
(previously when the "AJAX" container var was set to true).

This was needed because the resulting HREF was used in both HTML
(where the `&` needed to be escaped) and in JavaScript (where it
must not be escaped).

Instead, we add the `{ajax: 1}` data to the ajax request in the
JavaScript where appropriate. In this case, it is only inside the
`toggleWepD` (weapon show/hide) function.