ChessGame: set endDate member consistently
Make sure `endDate` is set any time the game ends. Previously, it
was not being set on checkmate, and so it took the next page load
to register the end of the game.

We also consider the game ended _at_ the end time, not just after
it (so that it is set correctly on the page where the game ends).