ChessGame: set lastMove member consistently
Previously, `lastMove` was only set when `getMoves` was called, and
would not be updated for a move made in this page. This caused the
highlighted last move to be updated on the next page load instead of
on the page where the move is made.

To fix this, we move setting `lastMove` into `createMove`, so that
it is set correctly where it is called in both `getMoves` and
`tryMove`.

We also make `createMove` a private method.