Fix chess ajax triggers
Chess games will now update correctly in real time with ajax.

Since chess moves create new links, we need it to update the
SmrSession (so that those links remain valid). We no longer
add the `&ajax=1` query list parameter, which means that
chess_move_processing.php will run with `USING_AJAX=false`.

We do, however, still have `AJAX_CONTAINER=true`, so with the
recent change to `Template::display`, we will still output
XML as desired.

With this change, we can now do a standard container forward
in chess_move_processing.php instead of the hack with the
`SmrSession::$lastSN` and `do_voodoo`. This page will not set
any template parameters, but will add any `MoveMessage` to
the forwarded container for display (and they no longer show
up as dialogs because nobody likes dialogs).

In ajax.js, we need to update the `sn` global var when we
submit a chess move, so that the ajax updates use the new
SN instead of the original one when the page loaded.

BUGFIX: The `submitMoveHREF` needs to be displayed using
`Template::addJavascriptForAjax` so that the SN used by the
chess moves is correctly updated.

NOTE: If we wanted to make sure the displayed HREF was consistent
with what is being updated in the SmrSession, we can use the
following in the submitMove success function (but it should not
be necessary since the page is "always available"):

history.replaceState({}, null, submitMoveHREF);