Remove shop_gadget*.php
These are non-functioning scripts for selling "gadgets", which are
a feature for ships (from perhaps v1.5?) that no longer exist.

While there are still a lot of references to them in the code, these
scripts in particular aren't very useful in their current state.
Should gadgets ever be re-implemented, these scripts would likely be
rewritten from scratch anyway, so we just remove them.