Initialize Weapons index in combat results array
In player vs. player combat, if a team with more than 1 player had
a player with no weapons, then when checking if assists should be
awarded, we would get a warning when trying to iterate over the
'Weapons' index of the results array. This was because this index
was not initialized, and thus undefined if the player had no weapons.

We could fix this by skipping the check if 'Weapons' was not set,
but it is more robust to initialize 'Weapons' as an empty array.