Rework concept of date-related game fields
* start_date -> join_time
* start_turns_date -> start_time
* end_date -> end_time

It was confusing that the "start date" wasn't actually when the
game started. This rename helps clarify what these times mean.

Now, we don't really need to report the join time, except in
cases where the game has not started and the join time is earlier
than the start time (which would likely only happen for special
games that need some setup in-game, like Draft games). Otherwise,
it is an implementation detail that the players don't need to
know about.

We also change the suffix from "date" to "time" to more clearly
indicate that the field is an (integer) time.