Improve how starting turns are awarded
We used to prevent players from moving after joining a game, but
before it started, by setting their turns to 0, and only updating
turns once the game started. This required a delicate handling of
the `last_turn_update` player field.

Instead, we immediately award the starting turns when the player
joins, but we simply prevent them from moving by triggering an
error in the "move processing" scripts if the game has not yet
started.

This has the following benefits:

* Simplifies SmrPlayer::updateTurns by removing the logic related
  to awarding starting turns.

* Removes complications in voting for free turns, where you had to
  carefully ensure that you couldn't rewind the `last_turn_update`
  to before the game's start time. Now, you can vote for free turns
  immediately once the game starts.

* Fixes issue with players having 0 turns until validated, which
  prevented new players from trying out the game (since they would
  always hit the validation check in SmrPlayer::updateTurns).

* Fixes not being able to move until your first turn is awarded
  (i.e. until a few minutes after the game has started). Now you can
  move immediately once the game starts.

NOTE: If you decide to change the start time after players join,
this will still cause an inconsistency, since for players who join
before the start time, we set `last_turn_update` to the start time.
So to be consistent, you'd still need to also update this field for
all players, but the stakes aren't as high, since it is no longer
tied to awarding starting turns as well.