Remove unnecessary SmrGame wrappers in Globals
Remove the following Globals methods:
* getGameDescription
* getGameSpeed
* getGameMaxTurns
* getGameMaxPlayers
* getGameCreditsRequired
* getGameIgnoreStats
* getAllianceMaxPlayers
* getAllianceMaxVets
* getStartingCredits

Some of these were unused. For those that were used in some places,
we can easily replace them with calling the methods directly on an
SmrGame instance that we can access.

We did need to add the following methods to more easily access an
SmrGame instance in some classes:
* SmrPlanet::getGame
* SmrAlliance::getGame
* SmrShip::getGame
* SmrPort::getGame

Also some minor refactoring of game_join_processing.php.