Add "Database Cleanup" admin tool
This replaces the functionality of the `cleanup_old_data.sql`
script in `db/maintenance`.

We also provide info about the size of the database, the number
of rows deleted from each table, and the change in database size.

Add `route_cache` (for NPC's) to the list of tables to clean up.