Remove album/search_processing.php
This page was responsible for a lot of log spam:

PHP Notice:  Undefined index: nick

To fix this, it is simplest to just remove the page and have the
search form go directly back to index.php.

Switch to using the `nick` query param as the search term instead
of the query param name itself. (We could allow both, but then if
a user makes their HoF name `nick=`, it would be displayed for an
empty search.)