Add ability to decline and abandon missions
When you can accept a mission, you now also have the option
to decline it, which will prevent it from displaying again.

Once you have accepted a mission, you can now also abandon
it at any time, which will allow you to restart it later.

Use a new `Mission` class to hold the HREF methods:
* getAcceptHREF (from Globals::getAcceptMissionHREF)
* getClaimRewardHREF (from Globals::getClaimMissionRewardHREF)
* getDeclineHREF (new)
* getAbandonHREF (new)

Add method `SmrPlayer::declineMission`, which uses a new "step"
argument to `addMission` to reuse that code to effectively add
the mission in its completed state.

Also remove the `SmrPlayer::getCompletedMissions` method because
it is unused and non-functional (it queries a non-existent table
`player_completed_mission`).