Add img dimensions to President icon
Specifying the image dimensions allows the browser to allocate
space for the image before the image is downloaded. This prevents
other page elements from being shifted after the image is loaded
(which makes the page flicker and look like it's jumping around).

This is certainly worth a small amount of additional text being
served.