SmrPlayer: modify allowed zoom levels
Previous: [0, 12]
New: [1, 9]

The Local Map does not make sense as a single sector, in part
because it disables all the hotkeys for movement, which is why the
new minimum zoom level is 1 (3x3) instead of 0 (1x1).

Reduce the maximum zoom level from 12 -> 9 simply because the
Local Map is not yet optimized and slows down a lot when displaying
many sectors.