Remove items related to unused layout
The repository contains remnants of a proposed change to the page
design, using `smr_layout.png` as the background image. This design
is not used, and will not be used in the foreseeable future.

We remove a handful of CSS id selectors from Default.css and a sample
image (`new_inter_basic.png`) showing what the background would look
like when in use.

We retain the original proposed background (`smr_layout.png`) for
historical reasons.