Change Local Map zooming
Replace the mouse wheel zooming with zoom expand/shrink buttons.

While the mouse wheel was a creative idea, it was perhaps not the
most ideal implementation for a few reasons:

* The explanation of how to use it took up prime real estate on
  the Local Map page. (Without it, the page height doesn't change
  between the Local Map and Current Sector.)

* The sensitivity of mouse wheels (especially the new continuous
  types that don't have discrete notches) combined with the delay
  in loading the page made it hard to use precisely. This also
  could result in "Multiple Actions" errors for users with slow
  connections to the server.

* Players are not changing the Local Map size often, so the dynamic
  nature of the mouse wheel solution was under-utilized.

In replacing the old system, we can remove a good chunk of non-cached
javascript that was being served with each page. Also removed a few
unused session variables in the zoom links.