forces_drop_processing.php: clean up inputs
Add a local function `getInputAmount` to handle the processing of
each input variable. We change the `create_error` for non-numeric
input to an Exception, since the HTML5 type restriction means that
non-numeric input implies falsified form data.

Since the output of `round` is numeric, we perform the `is_numeric`
check prior to rounding. This fixes the following two Scrutinizer
warnings (repeated for each input element):

* It seems like `IssetNode ? $var['drop_m...UEST['drop_mines']) : 0`
  can also be of type `string`; however, parameter `$val` of `round()`
  does only seem to accept double, maybe add an additional type check?

* The condition `is_numeric($drop_mines)` is always true.