Chess matches will now end with each game
Having chess matches span multiple games caused two major problems:

1. It was a sneaky way to keep track of who a player is from game
   to game, circumventing the anonymity that everyone assumed they
   had.

2. If one player in a match doesn't join a subsequent game, then
   we can get any number of Exceptions due to trying to access
   methods on a player that is `null`.

The simplest solution to these problems is to have matches end when
the game ends.

To implement this, we need to make a few changes:

* Add a `game_id` column to the `chess_game` table.
  (Since we can't know what game_id past chess games were associated
  with, they will be assigned `game_id=0`.)

* Replace the `ChessGame` method `getOngoingAccountGames` with
  `getOngoingPlayerGames`, which is restricted to matches in the
  current game.

* Let the `ChessGame` methods `get<X>Player` raise an Exception
  if the SmrPlayer isn't found, rather than returning null.
  (There should be no more valid cases where the player isn't found.)

Since it is no longer possible for a player to be `null`, we can
remove those conditionals in the chess.php template. We can also
clean up some of the `ChessGame` methods so that they don't use
references.

Also added some flavor text to the "Casino" page.